2-3: integer division
The close runner up for "most fun part": integer division

We leave in all ".0" which make no difference on py3 (nor on py2 with a
future import) but are necessary on py2, and we add in int() where
necessary to make py3 behave like py2.

Two minor "exceptions":
- We add in a comment where the code makes sure that we divide floats
  (so that py2 does not need an extra float() or .0) but where this is
  hard to spot.
- We let FormatTime convert its argument to int. Apparantly, the code
  expected to be fed an int on py2 but would break if not. So, it is
  more robust on py2 now and equivalent otherwise.



git-svn-id: http://svn.emphy.de/impressive/branches/python3/impressive@266 58003caa-664d-4c56-a378-6d187ff66ee1

+++ src/control.py
@@ -427,7 +427,7 @@ def PrepareTransitions():
     # STEP 1: randomly assign transitions where the user didn't specify them
     cnt = sum([1 for page in range(1, PageCount + 1) \
                if GetPageProp(page, 'transition', Unspecified) == Unspecified])
-    newtrans = ((cnt / len(AvailableTransitions) + 1) * AvailableTransitions)[:cnt]
+    newtrans = (int(cnt / len(AvailableTransitions) + 1) * AvailableTransitions)[:cnt]
     random.shuffle(newtrans)
     for page in range(1, PageCount + 1):
         if GetPageProp(page, 'transition', Unspecified) == Unspecified:
